Added -h/--help flag

+++ resources/modules/module_helpers/metadata_helper.py
@@ -16,8 +16,28 @@ import sys
 import os
 
 
-def run_mdh(sw_name, path=False):
+def run_mdh(sw_name, path=False, overwrite=False):
+	"""
+	The pythonSearcher.py driver file.
 
+	Parameters
+	----------
+	sw_name: str
+	Name of the software to be extracted
+
+	path: str
+	Path to the software stack root folder
+
+	overwrite: boolean
+	True/False flag. If overwrite is True, then jsons
+	are overwritten.
+
+	Returns 
+	-------
+	data: str
+	Returns the value of the variable.
+
+	"""
 	if path:
 		c_path = path
 	else:
@@ -31,10 +51,10 @@ def run_mdh(sw_name, path=False):
 		})
 
 	template_dict = {
-	      "CHANNEL_LINK": None,
-	      "NAME": sw_name,
-	      "INFO": None,
-	      "License": None
+			"CHANNEL_LINK": None,
+			"NAME": sw_name,
+			"INFO": None,
+			"License": None
 	}
 
 	bioconda_search = pySrch.searchBioconda()
@@ -55,6 +75,8 @@ def run_mdh(sw_name, path=False):
 		data = template_dict
 		not_found = True 
 
+	def_name = '.metadata.json'
+
 	f_name = '.metadata.json'
 
 	if pypi:
@@ -63,18 +85,35 @@ def run_mdh(sw_name, path=False):
 		f_name = 'NOTFOUND_' + f_name
 
 	json_path = os.path.join((c_path), sw_name)
+	check_json_path = os.path.join(json_path, def_name)
 	json_filepath = os.path.join(json_path, f_name)
 
 	if not os.path.exists(json_path):
 		os.makedirs(json_path)
 
-	with open(json_filepath, 'w') as f:
-		json.dump(data, f, indent=6)
+	if overwrite:
+		with open(json_filepath, 'w') as f:
+			json.dump(data, f, indent=6)
+		print(f'JSON inside {json_filepath}')
 
-	print(f'JSON inside {json_filepath}')
-
+	elif os.path.isfile(check_json_path):
+		print(f'JSON skipped since exists {check_json_path}')
+		pass
+	
 if __name__ == '__main__':
+	try:
+		sys.argv[1]
+	except Exception as IndexError:
+		print("Are you sure you're using it with proper arguments?\n \
+			Try using -h or --help for help".replace('		', ''))
+		sys.exit()
+	if sys.argv[1] == '-h' or '--help':
+		print("Driver Script to get a software's metadata\n\
+		Usage:\n\
+		python metadata_helper.py <SOFTWARE_NAME> <PATH_TO_SOFTSTACK> \
+		<OVERWRITE_FLAG=True/False(False by default)> ".replace('		', ''))
+		sys.exit()
 	sw_name = sys.argv[1]
 	if len(sys.argv) == 3:
 		path_ = sys.argv[2]
-	run_mdh(sw_name, path_)
+	run_mdh(sw_name, path=path_, overwrite=False)
